Documentation de l'API

+++ src/main/java/com/elprofesor/collaborationtool/server/services/TaskServiceJPA.java
@@ -1,16 +1,18 @@
 package com.elprofesor.collaborationtool.server.services;
 
 import com.elprofesor.collaborationtool.server.controllers.NotFoundException;
+import com.elprofesor.collaborationtool.server.entities.Project;
 import com.elprofesor.collaborationtool.server.entities.Task;
+import com.elprofesor.collaborationtool.server.entities.Users;
 import com.elprofesor.collaborationtool.server.mapper.TaskMapper;
 import com.elprofesor.collaborationtool.server.mapper.UserMapper;
-import com.elprofesor.collaborationtool.server.models.ProjectRequestDTO;
 import com.elprofesor.collaborationtool.server.models.TaskRequestDTO;
 import com.elprofesor.collaborationtool.server.models.TaskResponseDTO;
 import com.elprofesor.collaborationtool.server.repositories.ProjectRepository;
 import com.elprofesor.collaborationtool.server.repositories.TaskRepository;
 import com.elprofesor.collaborationtool.server.repositories.UserRepository;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -43,36 +45,59 @@ public class TaskServiceJPA implements TaskService {
     }
 
     @Override
-    public TaskResponseDTO saveNewTask(TaskRequestDTO taskRequestDTO) {
+    public TaskResponseDTO saveNewTask(UUID projectId, TaskRequestDTO taskRequestDTO, Users currentUser) {
+        Optional<Project> projet = projectRepository.findById(projectId);
         Task taskTosave = taskMapper.taskRequestDtoToTask(taskRequestDTO);
         taskTosave.setProject(projectRepository.findByTitleContainingIgnoreCase(taskRequestDTO.getProjectName()));
         taskTosave.setAssign_to(userRepository.findByUsername(taskRequestDTO.getAssign_to()).get());
-        return taskMapper.taskToTaskResponseDto(taskRepository.save(taskTosave));
-    }
-
-    @Override
-    public Optional<TaskRequestDTO> updateTask(UUID id, TaskRequestDTO taskRequestDTO) {
-
-        AtomicReference<Optional<TaskRequestDTO>> atomicReference = new AtomicReference<>();
-        taskRepository.findById(id).ifPresentOrElse(foundTask -> {
-            foundTask.setTitle(taskRequestDTO.getTitle());
-            foundTask.setStatus(taskRequestDTO.getStatus());
-            foundTask.setDescription(taskRequestDTO.getDescription());
-            foundTask.setAssign_to(userRepository.findByUsername(taskRequestDTO.getAssign_to()).get());
-            Task savedTask = taskRepository.save(foundTask);
-            atomicReference.set(Optional.of(taskMapper.taskToTaskRequestDto(savedTask)));
-        }, () -> {
-            atomicReference.set(Optional.empty());
-        });
-        return atomicReference.get();
-    }
-
-    @Override
-    public Boolean deleteTask(UUID id) {
-        if(taskRepository.existsById(id)){
-            taskRepository.deleteById(id);
-            return true;
+        if(projet.get().getOwner().equals(currentUser)){
+            return taskMapper.taskToTaskResponseDto(taskRepository.save(taskTosave));
+        }else{
+            throw new AccessDeniedException("Seul l'owner du projet peut ajouter une nouvelle tâche au projet");
         }
+
+    }
+
+    @Override
+    public Optional<TaskRequestDTO> updateTask(UUID id, TaskRequestDTO taskRequestDTO, Users currentUser) {
+        Optional<Task> tache = taskRepository.findById(id);
+        Project projet = projectRepository.findByTitleContainingIgnoreCase(tache.get().getTitle());
+        if(projet.getOwner().equals(currentUser)){
+            AtomicReference<Optional<TaskRequestDTO>> atomicReference = new AtomicReference<>();
+            taskRepository.findById(id).ifPresentOrElse(foundTask -> {
+                foundTask.setTitle(taskRequestDTO.getTitle());
+                foundTask.setStatus(taskRequestDTO.getStatus());
+                foundTask.setDescription(taskRequestDTO.getDescription());
+                foundTask.setAssign_to(userRepository.findByUsername(taskRequestDTO.getAssign_to()).get());
+                Task savedTask = taskRepository.save(foundTask);
+                atomicReference.set(Optional.of(taskMapper.taskToTaskRequestDto(savedTask)));
+            }, () -> {
+                atomicReference.set(Optional.empty());
+            });
+            return atomicReference.get();
+        }else{
+            throw new AccessDeniedException("Seul un owner peut modifier les informations d'une tâche.");
+        }
+
+    }
+
+    @Override
+    public Boolean deleteTask(UUID id, Users currentUser) {
+        Optional<Task> tache = taskRepository.findById(id);
+        System.out.println("Nom de la tâche trouvée : " + tache.get().getTitle());
+        Project projet = projectRepository.findByTitleContainingIgnoreCase(tache.get().getProject().getTitle());
+        //System.out.println("Projet correspondant : " + projet.toString());
+        System.out.println("Owner trouvé : " + projet.getOwner().getUsername());
+        System.out.println("Utilisateur actuellement connecté : " + currentUser.getUsername());
+        if(projet.getOwner().equals(currentUser)){
+            if(taskRepository.existsById(id)){
+                taskRepository.deleteById(id);
+                return true;
+            }
+        }else{
+            throw new AccessDeniedException("Seul le owner du projet peut supprimer cette tâche");
+        }
+
         return false;
     }
 }
